Refactor integrations into integrations/core and integrations/optional
