+++ integrations/core/papers/test_sigir2022.py
@@ -20,7 +20,7 @@ import multiprocessing
 import os
 import unittest
 
-from integrations.utils import clean_files, run_command, parse_score, parse_score_msmarco
+from integrations.core.utils import clean_files, run_command, parse_score, parse_score_msmarco
 
 
 class TestSIGIR2022(unittest.TestCase):
