+++ integrations/core/sparse/test_lucenesearcher_check_robust04.py
@@ -16,32 +16,32 @@
 
 import unittest
 
-from integrations.lucenesearcher_anserini_checker import LuceneSearcherAnseriniMatchChecker
+from integrations.core.lucenesearcher_anserini_checker import LuceneSearcherAnseriniMatchChecker
 from pyserini.search.lucene import LuceneSearcher
 
 
-class CheckSearchResultsAgainstAnseriniForCore17(unittest.TestCase):
+class CheckSearchResultsAgainstAnseriniForRobust04(unittest.TestCase):
     def setUp(self):
         # Make sure the required index is downloaded.
-        LuceneSearcher.from_prebuilt_index('nyt')
+        LuceneSearcher.from_prebuilt_index('disk45')
 
         self.checker = LuceneSearcherAnseriniMatchChecker(
-            index='nyt',
-            topics='tools/topics-and-qrels/topics.core17.txt',
-            pyserini_topics='core17',
-            qrels='tools/topics-and-qrels/qrels.core17.txt')
+            index='disk45',
+            topics='tools/topics-and-qrels/topics.robust04.txt',
+            pyserini_topics='robust04',
+            qrels='tools/topics-and-qrels/qrels.robust04.txt')
 
     def test_bm25(self):
-        self.assertTrue(self.checker.run('core17_bm25', '-bm25', '--bm25'))
+        self.assertTrue(self.checker.run('robust04_bm25', '-bm25', '--bm25'))
 
     def test_bm25_rm3(self):
-        self.assertTrue(self.checker.run('core17_bm25_rm3', '-bm25 -rm3', '--bm25 --rm3'))
+        self.assertTrue(self.checker.run('robust04_bm25_rm3', '-bm25 -rm3', '--bm25 --rm3'))
 
     def test_qld(self):
-        self.assertTrue(self.checker.run('core17_qld', '-qld', '--qld'))
+        self.assertTrue(self.checker.run('robust04_qld', '-qld', '--qld'))
 
     def test_qld_rm3(self):
-        self.assertTrue(self.checker.run('core17_qld_rm3', '-qld -rm3', '--qld --rm3'))
+        self.assertTrue(self.checker.run('robust04_qld_rm3', '-qld -rm3', '--qld --rm3'))
 
     def tearDown(self):
         pass
