+++ integrations/core/sparse/test_prebuilt_unicoil.py
@@ -19,7 +19,7 @@
 import os
 import unittest
 
-from integrations.utils import clean_files, run_command, parse_score
+from integrations.core.utils import clean_files, run_command, parse_score
 
 
 class TestSearchIntegration(unittest.TestCase):
