+++ integrations/optional/dense/test_ance.py
@@ -20,7 +20,7 @@ import multiprocessing
 import os
 import unittest
 
-from integrations.utils import clean_files, run_command, parse_score_qa, parse_score_msmarco
+from integrations.core.utils import clean_files, run_command, parse_score_qa, parse_score_msmarco
 from pyserini.encode import QueryEncoder
 from pyserini.search import get_topics
