+++ integrations/optional/dense/test_dpr.py
@@ -21,7 +21,7 @@ import multiprocessing
 import os
 import unittest
 
-from integrations.utils import clean_files, run_command, parse_score_qa
+from integrations.core.utils import clean_files, run_command, parse_score_qa
 from pyserini.encode import QueryEncoder
 from pyserini.search import get_topics
