+++ integrations/optional/dense/test_encode.py
